==> notebooks/03_Modelling.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc" style="margin-top: 1em;"><ul class="toc-item"><li><span><a href="#Introduction" data-toc-modified-id="Introduction-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Introduction</a></span></li><li><span><a href="#Setup" data-toc-modified-id="Setup-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Setup</a></span></li><li><span><a href="#Libraries" data-toc-modified-id="Libraries-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Libraries</a></span></li><li><span><a href="#Data" data-toc-modified-id="Data-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Data</a></span></li><li><span><a href="#Modelling" data-toc-modified-id="Modelling-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Modelling</a></span><ul class="toc-item"><li><span><a href="#Training" data-toc-modified-id="Training-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Training</a></span></li><li><span><a href="#Evaluation" data-toc-modified-id="Evaluation-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Evaluation</a></span></li></ul></li></ul></div>

# Introduction

This notebook trains an ALS model on the MovieLens 20M Dataset.

# Setup

%%capture
%cd ..

# Libraries

from pyspark.ml.recommendation import ALS
from pyspark.mllib.evaluation import RankingMetrics
import pyspark.sql.functions as f

from src import data

# Data

df = data.get_data("/tmp/ml-20m/ratings.csv")
df = data.add_label(df).cache()

train_df, valid_df, test_df = data.timesplit_data(df, train_dates=("2012-01-01", "2014-11-01"))

df.show(5)

Check date ranges in the train, valid and test splits:

for df_split in [train_df, valid_df, test_df]:
    df_split.selectExpr("count(timestamp)", "min(timestamp)", "max(timestamp)").show()

# Modelling

## Training

Specify model hyperparameters:

rank = 400
maxIter = 20
coldStartStrategy = "drop" # no difference if switch to "nan"!
implicitPrefs = False

Instantiate the model:

als = ALS(rank=rank, 
          maxIter=maxIter,  
          coldStartStrategy=coldStartStrategy, 
          implicitPrefs=implicitPrefs,
          userCol="userId", 
          itemCol="movieId", 
          ratingCol="adj_rating")

Train the model:

%%time
model = als.fit(train_df)

## Evaluation

Evaluate the model on the validation set using the NDCG metric:

recsize = 10

valid_df.show(5)

Build the predicted rankings for each user in the validation set:

valid_users = valid_df.select("userId").distinct()
predicted_ranks = model.recommendForUserSubset(valid_users, recsize)
# predicted_ranks.show(10)

We just need the list of movies for each user:

predicted_ranks = predicted_ranks \
    .rdd \
    .map(lambda r: (r.userId, [row.movieId for row in r.recommendations])) \
    .toDF(["userId", "predicted_ranking"])

Build the ground truth rankings:

actual_ranks = data.create_ground_truth_rankings(valid_df)

Check how bad is the cold start problem for new users:

# actual_ranks \
#     .join(predicted_ranks, "userId", "left") \
#     .withColumn("missing_recs", f.isnull("predicted_ranking").cast("int")) \
#     .selectExpr("count(missing_recs) AS total_users",
#                 "sum(missing_recs) AS new_users",
#                 "round(mean(missing_recs), 4) AS frac_users_with_missing_recs") \
#     .show(5)

Compute the metrics:

prediction_and_labels = predicted_ranks \
    .join(actual_ranks, "userId", "inner") \
    .drop("userId") \
    .rdd

metrics = RankingMetrics(prediction_and_labels)

metrics.ndcgAt(recsize)

metrics.precisionAt(recsize)
