==> notebooks/04_Dataset_Split.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc" style="margin-top: 1em;"><ul class="toc-item"><li><span><a href="#Introduction" data-toc-modified-id="Introduction-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Introduction</a></span></li><li><span><a href="#Setup" data-toc-modified-id="Setup-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Setup</a></span></li><li><span><a href="#Libraries" data-toc-modified-id="Libraries-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Libraries</a></span></li><li><span><a href="#Data" data-toc-modified-id="Data-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Data</a></span></li><li><span><a href="#Split" data-toc-modified-id="Split-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Split</a></span></li></ul></div>

# Introduction

This notebook splits the dataset into a train, validation and test split:

# Setup

%%capture
%cd ..

# Libraries

from src import data
import pyspark.sql.functions as f
from lenskit import crossfold as xf

from pyspark.sql import SparkSession

# Data

df = data.get_data("/tmp/ml-20m/ratings.csv")
df.show(5)

df = df \
    .withColumnRenamed("userId", "user") \
    .withColumnRenamed("movieId", "item") \
    .cache()
df.show(5)

# Split

Check timestamp range:

df.selectExpr("min(timestamp)", "max(timestamp)").show()

Training on whole dataset may take too long so we filter by timestamp to reduce dataset size:

df = df.filter(f.expr("timestamp >= '2010-01-01'"))
df.show(5)

%%time
train_df, test_df = next(xf.partition_users(df.toPandas(), 1, xf.SampleFrac(0.2)))

nrows_train, _ = train_df.shape
nrows_test, _ = test_df.shape

print(f"Number of rows in training set: {nrows_train:,d}")
print(f"Number of rows in test set: {nrows_test:,d}")

Save the results:

 spark = SparkSession.builder.getOrCreate()

%%time
# df.write.format(source).mode("overwrite").save(path)
spark.createDataFrame(train_df).write.mode("overwrite").parquet("/tmp/ml-20m/train_df.parquet")
spark.createDataFrame(test_df).write.mode("overwrite").parquet("/tmp/ml-20m/test_df.parquet")
