==> notebooks/08_Evaluate_Reranking_Impact.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc" style="margin-top: 1em;"><ul class="toc-item"><li><span><a href="#Introduction" data-toc-modified-id="Introduction-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Introduction</a></span></li><li><span><a href="#Setup" data-toc-modified-id="Setup-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Setup</a></span></li><li><span><a href="#Libraries" data-toc-modified-id="Libraries-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Libraries</a></span></li><li><span><a href="#Data" data-toc-modified-id="Data-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Data</a></span></li><li><span><a href="#Evaluation" data-toc-modified-id="Evaluation-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Evaluation</a></span><ul class="toc-item"><li><span><a href="#Without-Reranking" data-toc-modified-id="Without-Reranking-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Without Reranking</a></span></li><li><span><a href="#With-Reranking" data-toc-modified-id="With-Reranking-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>With Reranking</a></span></li></ul></li></ul></div>

# Introduction

This notebook evaluates the impact of the reranking algorithm on NDCG and coverage of longtail items. 

# Setup

%%capture
%cd ..

# Libraries

import pandas as pd

from typing import List
from pyspark.sql.types import Row

from src import preference_reranker as pr

from lenskit import topn

from pyspark.sql import SparkSession
from pyspark.sql import functions as F

# Data

spark = SparkSession.builder.getOrCreate()

test_df = spark.read.parquet("/tmp/ml-20m/test_df.parquet/").toPandas()
test_df.head(5)

eval_df = spark.read.parquet("/tmp/ml-20m/evaluation_dataset.parquet") \
    .orderBy("user", "rank")
eval_df.show(5)

eval_df.selectExpr("count(DISTINCT user)", "min(rank)", "max(rank)").show()

movie_cat = spark.read.parquet("/tmp/ml-20m/movie_categories.parquet") \
  .withColumnRenamed("movieId", "item") \
  .withColumnRenamed("category", "item_category")

movie_cat.show(5)

user_pref = spark.read.parquet("/tmp/ml-20m/user_preference.parquet") \
      .withColumnRenamed("userId", "user") 
user_pref.show(5)

# Evaluation

Setup to calculate NDCG:

rla = topn.RecListAnalysis()
rla.add_metric(topn.ndcg)

## Without Reranking

NDCG:

df = eval_df.filter("rank <= 10").toPandas() 
print(f"NDCG@10: {rla.compute(df, test_df).ndcg.mean():.4f}")

Longtail Coverage:

eval_df \
    .filter("rank <= 10") \
    .join(movie_cat, "item", "left") \
    .groupBy("item_category") \
    .agg(F.expr("count(item) AS total_items"),
         F.expr("count(DISTINCT item) AS unique_items")) \
    .show(truncate=False)

## With Reranking

For each user, construct a candidate set and longtail preference:

user_details = eval_df \
    .join(movie_cat, "item", "left") \
    .withColumn("movie", F.struct(F.col("item").alias("movie_id"),
                                  F.col("score").alias("base_score"),
                                  F.col("item_category").alias("category"))) \
    .groupBy("user") \
    .agg(F.expr("collect_set(movie) AS candidate_set")) \
    .join(user_pref, "user", "left") \
    .toPandas()

user_details.head(5)

Loop through each row to construct a reranked recommendation for each user:

recsize = 10
longtail_weight = 0.8

def pyspark_rows_to_movies(rows: List[Row]):
    return [pr.Movie(row.movie_id, row.base_score, row.category) 
            for row in rows]

reranked_recs = []

for user, candidate_set, longtail_pref in user_details.itertuples(name=None, index=False):
    candidate_set = pyspark_rows_to_movies(candidate_set)
    new_recs = pr.construct_reclist(candidate_set=candidate_set,
                                    size=recsize,
                                    longtail_pref=longtail_pref,
                                    longtail_weight=longtail_weight)
    
    reranked_recs.append(new_recs)

user_details["reranked_recs"] = reranked_recs

user_details.head(5)

Transform `user_details` into `eval_df` form:

df = user_details[["user", "reranked_recs"]] \
    .explode("reranked_recs")

df["item"] = df["reranked_recs"].apply(lambda m: m.movie_id)
df["score"] = df["reranked_recs"].apply(lambda m: m.base_score)

df = df[["item", "score", "user"]]
df["rank"] = df.groupby("user")["score"].rank("dense", ascending=True)

df = df.sort_values(by=["user", "rank"])

df.head(5)

NDCG:

print(f"NDCG@10: {rla.compute(df, test_df).ndcg.mean():.4f}")

Longtail Coverage:

df1 = df.merge(movie_cat.toPandas(), 
               on="item", how="left") \
  .groupby("item_category") \
  .size() \
  .to_frame("total_items")

df2 = df.merge(movie_cat.toPandas(), 
               on="item", how="left") \
  .groupby("item_category") \
  .item \
  .nunique() \
  .to_frame("unique_items")

pd.concat([df1, df2], axis=1)
